+++ lib/customerio.rb
@@ -5,6 +5,7 @@ module Customerio
   require "customerio/base_client"
   require "customerio/client"
   require "customerio/requests/send_email_request"
+  require "customerio/requests/send_push_request"
   require "customerio/api"
   require "customerio/param_encoder"
 end